resource-timing/iframe-srcdoc-allow-sameorigin-resources.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>iframe srcdoc resources not exposed to parent page</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
  var img_entries = 0;
  var img_url = "resources/blue.png";
  var observer = new PerformanceObserver(
    function (entryList, obs) {
      var entries = entryList.getEntriesByType("resource");
      for (var i = 0; i < entries.length; ++i) {
        if (entries[i].name.indexOf(img_url) != -1)
            ++img_entries;
      }
    });
  observer.observe({entryTypes: ["resource"]});
  window.onload = function() {
    // A timeout is needed as PerformanceObserver is not guaranteed to run before onload triggered.
    setTimeout(function() {
      test(function(){
        assert_equals(img_entries, 0);
      }, "Only one resource entry per resource");
    },0)
  };
  // Images are added dynamically to make sure the observer is registered before their download finishes.
  var iframe = document.createElement("iframe");
  iframe.srcdoc = "<img src='resources/blue.png'>";
  iframe.sandbox = "allow-same-origin";
  document.body.appendChild(iframe);
</script>
<h1>iframe srcdoc resources not exposed to parent page</h1>
<div id="log"></div>
</body>
